Remove locking and check for double events.
We are observing behavior in some resizing that content loads, tells the
parent to resize the iframe and locks it. Then, more content loads, but
since the trigger is locked, it will not appropriately resize the outer
frame.

This removes the lock and the double event check, because it is
sufficient enough to only run the resizing when the content has changed
size.

Another option is to remember the size as it changes, and send it up to
the parent as soon as it unlocks, but this is simpler.

It looks like this has been in the file for some time; if there is
another reason for having these extra checks, maybe we haven't run into
it.