Allow eventCancelTimer to be configurable
With simple height calculations, we don't need to worry about the
overhead of working out the size of the content, and thus it is best to
not lock the communication between iframes, and allow the height to
always be updated to reflect the most up-to-date content height.

This sets the default `eventCancelTimer` to zero, which then skips the
logic for locking the trigger and allows it to update immediately.

The readme is also updated to reflect this new option and make clear
when these settings should be changed from the default.