Issue #244 iframe not resizing correctly when images are included in dynamically loaded content
Added back support for checking each mutation for images.  Replaced
checked on image height/width with check on image.complete flag as
height/width seem to be set in IE and Firefox with erroneous values,
even if not set on the actual image tag. Replaced bind approach for
registering image load event listener with actual event listener
function to avoid registering multiple listeners for same image and also
so the listener can be removed.  Also added listener for error event so
that images which fail to load are handled cleanly.