Correct import statement and typos
The example may work as written, but the directive would be relying on a global `window.iFrameResize` function, not the one imported, `iframeResize`. If the global function isn't present or is shadowed by another script, the directive will fail or call the wrong function.

Furthermore, importing at the package root will include the contents of both `js/iframeResize.js` and `js/iframeResizer.contentWindow.js`. Module bundlers like webpack can't tree-shake the files because of side effects.  Better to just import the default export of `js/iframeResizer.js` if that's all we're going to use.